requirements.txt
aiokafka==0.10.0
PyYAML==6.0.2
websockets==12.0

